Update Program.cs
In line 110, if the type of position is decimal, the "MoveTo" method will return an error, saying that "No encoder defined". Change the type to int can solve the problem.

+++ C#/KCube/KIM101/Program.cs
@@ -107,7 +107,7 @@ namespace KIM_Console_net_managed
 
             // Optionally set the position
 
-            decimal position = 0m;
+            int position = 0m;
             if (position > 0m)
             {
                 try
